Added Geological Phenomena to Caves
- Collectable.cs found in Assets/Scripts/Caves/Collectable.cs contains the script for the collectable items: rock and geological phenomena. Inside the script their values are initialized and applied to the inventory UI (Joseph will take care of applying it to the inventory UI)
- There are numerous rocks in the cave scenes and a total of 35 geological phenomena to count towards the skill tree
