+++ Assets/Animations/RegularText.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 88c6bcb79a453c143a1b469ec8ab5b60
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
