+++ Assets/Animations/RegularTextUp.anim.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 37a52a2627e6b194c88ab895afab01a5
-DefaultImporter:
+guid: 1446d988d3e603646a0e04ea5fea00e8
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 7400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
