+++ Assets/Scripts/Caves/Collectable.cs.meta
