+++ ProjectSettings/TagManager.asset
@@ -14,6 +14,7 @@ TagManager:
   - ReminderText
   - RobotBuddy
   - Rock
+  - GeologicalPhenomena
   layers:
   - Default
   - TransparentFX
